Fix relations w/ non-fluent query methods
